drag outside slider fix
